Add (basic, unoptimized) support for deleting data.
Add tests to make sure deletes work with and without VACUUM. Under
the hood the implementation actually makes no changes to the index
itself, instead relying on the indexscan to ignore tuples marked
as deleted by the heap.

In the future we should optimize this inside of ambulkdelete.

Unclear if this will work with index-only scans but those aren't
supported for now anyway...`