Fix handling of many graph nodes at the same location
When nodes have the same location, their distance to each other is 0.
And, their distance to all their neighbors is the same. In the case
with few maximum neighbors this can cause problems because the nodes
end up linking to each other and getting cut off from the rest of
the graph.

We fix this by introducing a secondary "tie break" distance between
nodes at the same location (based on on-disk distance). That's
used to create a secondary ordering for equivalent nodes and allows
the prune function to then properly cut away equivalent nodes so
that your neighbor list doesn't get full of them.

Note with SBQ, equivalence classes are actually not /so/ rare especially
with low dimension counts.