Fix HOT update handling
Previously we weren't handling HOT chain correctly because we used
fetch_row_version function instead of the index_fetch_tuple function
which correctly handles HOT. This commit changes that.

We also fix the snapshot that is used to be aligned with the snapshot
of the scan, which is more correct.

Finally we handle the case that no valid heap tuple is found. While
we don't quite know how that can be hit, it's better to handle that
case.