Avoid creating a new page on every insert (#148)
Currently, `insert_storage` in the access method creates a new Tape
to add the vector to, on every insert. This can cause a lot of empty
space in pages. This commit allows tapes to be resumed at the most
recent block (of the right tape), instead of always starting at a new
block. This allows `insert_storage` to continue writing into an existing
page instead of creating a new one.

Signed-off-by: Smitty <smitty@timescale.com>