Fit and finish before 0.5.0 release PR (#161)
Just to keep the actual deployment PR simpler, some fit and finish
changes:
* Updated README.md to cover L2 distance support
* Tweaks to upgrade tests to (a) account for the possibility of upgrade
from 0.4.0 on pg17, and (b) mark upgrade tests as serial to get us
closer to allow running them all in one go. There is still an issue with
pgrx dropping the test schema that needs to be fixed to get this
actually working, but we can deal with that later.