Only normalize vectors if index uses cosine distance (#165)
## Description

Cosine vector indices benefit from pre-normalization of vectors when
added to index, and one-time normalization of the query vector, so that
distance computation reduces to a simple inner product. L2 and other
metrics, however, are sensitive to vector magnitudes, so we should not
normalize in that case.

## Testing

Added a unit test that fails without my fix.