Have cargo-pgrx install command specify version (#187)
Right now if you run the command to install cargo-pgrx from the README
or DEVELOPMENT file, it doesn't work because it tries to install the
latest version, while we require an older patch version. This PR changes
these files to have the installation command read the required
cargo-pgrx version from the `Cargo.toml` file.

I didn't write the actual pgrx version (currently 0.12.5) in the file to
avoid adding another thing that needs to be changed every pgrx version
bump.