Seperate quantizer training from index build (#235)
This splits out quantizer training from `do_heap_scan`. This doesn't
change any of the behaviour of index-building, but is going to be useful
in a future PR that adds support for parallel indexing.

This PR also makes it so the creation progress is set properly when
building a plain (non-quantized) index.